requirements.txt
# Azure OpenAI + FastAPI (kyobo2 참고)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
openai>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.23.0
requests>=2.28.0
google-auth>=2.0.0
motor>=3.0.0
pymongo>=4.0.0
PyJWT>=2.8.0
